files have been separated and calling main.cc works, but it can be further simplified
